pyproject.toml
[tool.poetry]
name = "turk-instructions"
version = "0.1.0"
description = ""
authors = ["klxu03 <beaver2025@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "4.12.0"
gitpython = "3.1.31"
beautifulsoup4 = "4.11.2"
requests = "2.28.2"
pandas = "1.5.3"
rouge-score = "0.1.2"
python-dateutil = "2.8.2"
colorama = "0.4.6"
pillow = "9.4.0"
transformers = "4.26.1"
tqdm = "4.64.1"
numpy = "1.24.1"
boto3 = "1.28.34"
jsonlines = "3.1.0"
django = ">=3.2.16,<3.3.0"
django-admin-autocomplete-list-filter = "1.0.1"
django-guardian = "2.4.0"
humanfriendly = "^10.0"
jsonfield = "3.1.0"
xlib = "^0.21"
pyautogui = "^0.9.54"
openai = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
